+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/ChatComponentFactory.kt
@@ -45,6 +45,7 @@ import io.getstream.chat.android.compose.ui.components.NetworkLoadingIndicator
 import io.getstream.chat.android.compose.ui.components.composer.ComposerLinkPreview
 import io.getstream.chat.android.compose.ui.components.composer.CoolDownIndicator
 import io.getstream.chat.android.compose.ui.components.composer.MessageInputOptions
+import io.getstream.chat.android.compose.ui.components.channels.MessageReadStatusIcon
 import io.getstream.chat.android.compose.ui.components.messages.DefaultMessageContent
 import io.getstream.chat.android.compose.ui.components.messages.DefaultMessageDeletedContent
 import io.getstream.chat.android.compose.ui.components.messages.DefaultMessageGiphyContent
@@ -906,6 +907,24 @@ public interface ChatComponentFactory {
         messageItem: MessageItemState,
     ) {
         MessageFooter(messageItem = messageItem)
+    }
+
+    /**
+     * The default status indicator of the message footer, weather the message is sent, pending or read.
+     */
+    @Composable
+    public fun MessageFooterStatusIndicator(
+        modifier: Modifier,
+        message: Message,
+        isMessageRead: Boolean,
+        readCount: Int,
+    ) {
+        MessageReadStatusIcon(
+            modifier = modifier,
+            message = message,
+            isMessageRead = isMessageRead,
+            readCount = readCount,
+        )
     }
 
     /**