+++ CHANGELOG.md
@@ -76,6 +76,7 @@
 - Create snapshot tests for messages stateless components. [#5578](https://github.com/GetStream/stream-chat-android/pull/5578)
 - Create message stateless component factories and refactor to have a single component factory object. [#5593](https://github.com/GetStream/stream-chat-android/pull/5593)
 - Create message list header component factories. [#5599](https://github.com/GetStream/stream-chat-android/pull/5599)
+- Create message footer status indicator stateless component factories. [#5598](https://github.com/GetStream/stream-chat-android/pull/5598)
 - Create message composer stateless component factories. [#5579](https://github.com/GetStream/stream-chat-android/pull/5597)
 
 ### ✅ Added